Various improvements to list.rs
I added the "Remove Obsolete Packages" entry to the list.rs file so that it can be easly executed. I also modified the other entries a bit so that they don't rely on bash to directly run the output of curl but instead pipe it through sh for cross-shell compatibility.