BugFix: Boot filesystem syntax fixes
Changes all instances of mkfs.vfat to mkfs.fat to prevent printed errors even if it should be a symlink.  Remove -t vfat from boot partition mount as it was reading out that it was not being recognized.  Changes mount point to /boot as opposed to /boot/efi to prevent chance of errors from mount operation.