Fix bluetooth manager (#959)
* Fix bluetooth errors

The i variable is edited locally as pipe is used. It results in error
when connecting to a device. The read -r at the last is not working for
some reason so added the read -r for each if case. Removed unused action
variable.

* Add bluez package to arch linux

* Remove unnecessary variable usage

We don't need to use device_list variable while devices variable can be directly used

* fix(bluetooth-manager): wrong display of device list