BugFix/Refactor: Various fixes for server-setup.sh (#1050)
* BugFix: ISO Code Sourcing

Changed the curl target to source the ISO country code, as the prior URL would output junk values.  These junk values breaks the intended functionality of reflector, pacman, and pacstrap, and can, under some circumstances, result in a segmentation fault.

* BugFix: Boot filesystem syntax fixes

Changes all instances of mkfs.vfat to mkfs.fat to prevent printed errors even if it should be a symlink.  Remove -t vfat from boot partition mount as it was reading out that it was not being recognized.  Changes mount point to /boot as opposed to /boot/efi to prevent chance of errors from mount operation.

* Refact: Various Fixes

1) Changes dhclient to dhcpcd for consistency.
2) Removes all systemctl --stop instances; stop is not available in chroot
3) Corrects value grepped from /proc/cpuinfo
4) Enables reflector systemd timer to start from boot.