docs: clean up issue templates and introduce new labels
heyy so i decided to do a little autumn cleaning in our issue management
in linutil. i created a couple of new labels
([label list](https://github.com/ChrisTitusTech/linutil/labels)),
updated the issue templates to be more concise, and updated
`ALLOWED_USERS` in the `issue-slash-commands` workflow.

from now on, we'll be using the new priority labels for issues and PRs.
also, i'd like to say that we (the maintainers, so @ChrisTitusTech,
@afonsofrancof, and me) need to start labeling stuff accordingly again.
the releases still rely on PR labels.
for that, i created the `initial review required` label. it'll be added
to every issue upon creation, and then removed after a maintainer review
which would include adding priority and scope labels.

so chris, please start labeling RPs again :3