feat(ci): automatic PR labeling
so i decided to make labeling prs a bit easier and this workflow
automatically syncs issue labels to a newly created pr
(if it has a closing keyword for that issue of course). this WILL NOT
always be accurate so we still need to review manually. it's not a
replacement, just makes stuff easier.

fully tested on my fork